Fix stale data flash by setting loading=true onDestroy of layout of workflow run
