+++ src/lib/components/event/event-summary.svelte
@@ -16,6 +16,7 @@
   function handleExpandChange(event: CustomEvent) {
     $expandAllEvents = event.detail.expanded;
   }
+  console.log('EVENT SUMMARY');
 </script>
 
 {#if loading}
