+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/__layout@root.svelte
@@ -1,0 +1,10 @@
+<script lang="ts">
+  import { onDestroy } from 'svelte';
+  import { loading } from '$lib/stores/workflow-run';
+
+  onDestroy(() => {
+    $loading = true;
+  });
+</script>
+
+<slot />
