+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/stack-trace.svelte
